239 Removed navigation properties: Nop.Core.Domain.Forums

+++ src/Presentation/Nop.Web/Controllers/BoardsController.cs
@@ -512,7 +512,8 @@ namespace Nop.Web.Controllers
 
             if (forumTopic == null)
                 return RedirectToRoute("Boards");
-            var forum = forumTopic.Forum;
+
+            var forum = _forumService.GetForumById(forumTopic.ForumId);
             if (forum == null)
                 return RedirectToRoute("Boards");
 
@@ -632,26 +633,28 @@ namespace Nop.Web.Controllers
                 });
 
             var forumPost = _forumService.GetPostById(id);
-            if (forumPost != null)
+
+            //TODO: ???? get topic one more time because it can be deleted (first or only post deleted)
+            var forumTopic = _forumService.GetTopicById(forumPost.TopicId);
+
+            if (forumPost != null && forumTopic != null)
             {
                 if (!_forumService.IsCustomerAllowedToDeletePost(_workContext.CurrentCustomer, forumPost))
                     return Challenge();
 
-                var forumTopic = forumPost.ForumTopic;
-                var forumId = forumTopic.Forum.Id;
-                var forumSlug = _forumService.GetForumSeName(forumTopic.Forum);
+                var forum = _forumService.GetForumById(forumTopic.ForumId);
+                var forumSlug = _forumService.GetForumSeName(forum);
 
-                _forumService.DeletePost(forumPost);
-
-                //get topic one more time because it can be deleted (first or only post deleted)
-                forumTopic = _forumService.GetTopicById(forumPost.TopicId);
                 if (forumTopic == null)
                 {
                     return Json(new
                     {
-                        redirect = Url.RouteUrl("ForumSlug", new { id = forumId, slug = forumSlug }),
+                        redirect = Url.RouteUrl("ForumSlug", new { id = forum.Id, slug = forumSlug }),
                     });
-                }
+                }                
+
+                _forumService.DeletePost(forumPost);
+                                
                 return Json(new
                 {
                     redirect = Url.RouteUrl("TopicSlug", new { id = forumTopic.Id, slug = _forumService.GetTopicSeName(forumTopic) }),
@@ -759,9 +762,9 @@ namespace Nop.Web.Controllers
                     var pageIndex = (_forumService.CalculateTopicPageIndex(forumPost.TopicId, pageSize, forumPost.Id) + 1);
                     var url = string.Empty;
                     if (pageIndex > 1)
-                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic), pageNumber = pageIndex });
+                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic), pageNumber = pageIndex });
                     else
-                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic) });
+                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic) });
                     return Redirect($"{url}#{forumPost.Id}");
                 }
                 catch (Exception ex)
@@ -807,11 +810,11 @@ namespace Nop.Web.Controllers
             if (!_forumService.IsCustomerAllowedToEditPost(_workContext.CurrentCustomer, forumPost))
                 return Challenge();
 
-            var forumTopic = forumPost.ForumTopic;
+            var forumTopic = _forumService.GetTopicById(forumPost.TopicId);
             if (forumTopic == null)
                 return RedirectToRoute("Boards");
 
-            var forum = forumTopic.Forum;
+            var forum = _forumService.GetForumById(forumTopic.ForumId);
             if (forum == null)
                 return RedirectToRoute("Boards");
 
@@ -871,11 +874,11 @@ namespace Nop.Web.Controllers
                     var url = string.Empty;
                     if (pageIndex > 1)
                     {
-                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic), pageNumber = pageIndex });
+                        url = Url.RouteUrl("TopicSlugPaged", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic), pageNumber = pageIndex });
                     }
                     else
                     {
-                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumPost.ForumTopic) });
+                        url = Url.RouteUrl("TopicSlug", new { id = forumPost.TopicId, slug = _forumService.GetTopicSeName(forumTopic) });
                     }
                     return Redirect($"{url}#{forumPost.Id}");
                 }
